Fixing bug for Repeated Courses
Fixes #2
To fix that bug, i am using a map from course name to the <li> element for
each course. whenever a new <li> for the same course name is encountered,
only the latter is stored in the map. Finally only elements in map have
their checkboxes checked by default.