Fixed bug for Repeated Courses (#4)
* Fixing bug for Repeated Courses

Fixes #2
To fix that bug, i am using a map from course name to the <li> element for
each course. whenever a new <li> for the same course name is encountered,
only the latter is stored in the map. Finally only elements in map have
their checkboxes checked by default.

* previous commit was other way around

Fixes #2

* Add an unchecked checkbox if course is already counted

Instead of totally ignoring it.

* Fixed indentation issues

* Implemented review comments

Using a set instead of map, storing and comparing course IDs instead of course names, separate function to append checkbox to an element, and better variable names consistent with the snake_case

* Fixed indentation